requirements.txt
# Core Django and web framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.7

# Environment and configuration
python-decouple==3.8

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Caching and background tasks
redis==5.0.1
celery==5.3.4

# File handling and utilities
Pillow==10.1.0
openpyxl==3.1.2

# API documentation
drf-spectacular==0.26.5

# Development and testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Production server
gunicorn==21.2.0

# Monitoring and logging
sentry-sdk==1.38.0

# Security
django-ratelimit==4.1.0

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# Excel and CSV processing
xlrd==2.0.1
csv-reader==1.2.0

# Mathematical calculations
sympy==1.12

# HTTP requests
requests==2.31.0

# JSON Web Tokens (if needed for advanced auth)
PyJWT==2.8.0

# Environment variables
python-dotenv==1.0.0